requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.1
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
isort==5.13.2
packaging==24.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
